requirements.txt
PyQt6>=6.6.0
Pillow>=10.0.0
pytest>=7.0.0
